Initial build of Nix Port based on Efl.
This adds the Nix Port infrastructure that builds
only WebCore, JSC and WebKit2. API, API Tests,
WebKitTestRunner and so on were left out for now.

WebKit1 is not supported.

This port is still strongly based on EFL and
has tons of cpp and headers files copied. A few
stubs were already added to WebCore/platform/nix
and to WebKit2/API/C/nix, but this was only the necessary
initial work to get it building as a new library.