Finishing clean up of WebKit2
We have added our own stub for LayerTreeContext, using a uint32 as layer id.
Both of our ProcessExecutablePath and ProcessLauncher we use are the GTK ones.
This also means that our WebProcess binary name is set in WebKit2/CMakeLists.txt
and is the same for Gtk, Efl and Nix (WebProcess), instead of setting it directly
in ProcessExecutablePath as Qt is doing.

It seems that after this patch Efl is only a memory for us.

Reviewed by: Caio Marcelo de Oliveira Filho