Implementing TestWebKitAPI for Nix
Now we can run WebKit2 API tests with Nix.
In order to test you need to pass
--cmakeargs="-DSHARED_CORE=1" to Tools/Scripts/build-webkit
and you can run the tests with Tools/Scripts/run-nix-tests.

A few tests are skipped in Tools/TestWebKitAPI/PlatformNix.cmake
and they should provide a good "ToDo" list of basic functions
and WebKit2 features we are still missing in Nix.

Reviewed by: Caio Marcelo de Oliveira Filho