Build with -DSHARED_CORE=1 by default
Now that the tests' infrastructure is being built
by default, we always need to build WebCore as a
shared library instead of a static one. This is a separate
issue, which is also present on the Efl port, that will
be solved later. For the time being enabling SHARED_CORE
fixes the build.

Reviewed by: Caio Marcelo de Oliveira Filho