Add support for Wheel events to WebView and MiniBrowser
One important change here is that we now keep track of the scroll
position in the WebView.

Wheel events that come from WebView API are fed into
WebPageProxy. Then they are sent to the WebProcess and processed by
WebCore code. If WebCore concludes we actually have a scroll to
happen, this information will propagate back to us via PageClient. It's
up to the WebView to handle this "callback" and set the visible content
rect accordingly.

Reviewed by: Jesus Sanchez-Palencia