WebView should use a proper structure to handle key events
Right now, the client should supply that event with proper enum
as keycode, and then internally WebView will figure out based
on that and on modifiers which string to pass to webkit keyboard
event (as text and unmodifiedText, see WebEvent.h) and the other
information needed.

We handle upper and lower case just checking whether we have shift
pressed while we press a letter. In the future we will need to handle
more cases (Caps Lock + Shift, for example), or even more symbols we
could get from a single key on keyboard (in case of different keyboard
formats, like pt-br which may have 3 different keys at once). Right now
we support at most 2 symbols on a key, and shift is the modifier that
triggers the change on it.

Reviewed by: Caio Marcelo
             Hugo Parente Lima