Unify the way to list feature default values on Nix port.
Currently there are at least two places to list and toggle webkit features, some features are listed in one place but not in the other and vice versa, the known places are:
Tools/Scripts/webkitperl/FeatureList.pm (used by build-webkit script) and cmake/WebKitFeatures.cmake, qt now use it own method to get the feature list, so I believe it doesn't use
FeatureList.pm default values anymore (please correct me if I'm wrong).

Some examples features in one place but not in the other:

ENABLE_CSS_IMAGE_SET, ENABLE_LLINT, ENABLE_MEMORY_SAMPLER, ENABLE_GLIB_SUPPORT, ENABLE_LEGACY_VIEWPORT_ADAPTION is on cmake but not on FeatureList.pm

Also it's a bit confusing when you need to add a new feature flag, because you need to know all places to add this feature, it's redundant and error prone.

Other problem (with cmake based ports) is if you set the default value of a flag on cmake but this flag is listed on FeatureList.pm, the default value on FeatureList.pm will overwrite
the cmake one, in other words when adding a feature you need to set the default value on both, cmake and FeatureList.pm, and keep them in sync.

My proposal is to have just two files, one listing all possible webkit features and other per-port listing the default (to true) features for this port, then FeatureList.pm and the
cmake files create the feature list and their default values based on this two files. No sync, no forgotten feature, no mess.

This commit solves the problem *JUST* for Nix, I'll try to push the commit upstream because I think it's very useful =], so comments, suggestions, etc before I push it to review
upstream are welcome.

Reviewed by: Caio Marcelo, Jesus Sanches-Palencia