Implement GLContextNix class and fixes MiniBrowser
This is needed for the current implementation TextureMapperGL that now
uses GraphicsContext3D to reuse GL code already available in WebKit.
In summary: MiniBrowser (GLX) is back to working, we need more work to
enable it properly so that other OpenGL based features work correctly.

The problem is that Cairo implementation of GraphicsContext3D relies
on WebCore::GLContext abstraction, a thin wrapper over OpenGL's
GLContext. The code path we were getting (RenderToCurrentGLContext) to
wasn't creating one for us.

On top of this, we can't really reuse the current GLX implementation
of it, since Nix can run in environments without GLX.

For now, GLContextNix will do nothing, since we only use it in the
UIProcess and the current context is never changed once we set. This
will be needed once we use offscreen contexts (like WebGL).

Given the wrapper is thin, the plan is to allow Nix applications to
customize it (via Platform hooks), as well as allowing creating new
offscreen buffers (needed for WebGL), and providing a proper pointer
to GetProcAddr function.

Reviewed by: Jesus Sanchez-Palencia