Create a GLContextEGL
Instead of relying on our own abstraction of GLContext and expose this to API as a recent
comment suggested, let's use EGL. It does pretty much what we needed: representation of a
GL context, a way to make it current, an implementation of getProcAddress(). It also will
handle creation of offscreen buffers in the future.

We needed to workaround CMake buildsystem to avoid getting EGL headers from inside
ANGLE. A better solution would be build ANGLESupport in a different directory, because
INCLUDE_DIRECTORIES have "per directory" scope.

Reviewed by Hugo Parente Lima and Noam Rosenthal