Change our MiniBrowser to use EGL instead of GLX
This commit also reorganize a bit initialization code of LinuxWindow
and get rid of some members that we don't need to store.

NOTE: The MiniBrowser needs EGL for testing on Linux desktop. In
general this works better with the opensource video drivers instead of
the proprietary versions. I'm testing on Ubuntu 12.04 and Nouveau
video driver.

Reviewed by Jesus Sanchez-Palencia and Noam Rosenthal.