Move our API tests to Tools/TestWebKitAPI
It is a more convenient place to put the tests, since we can share
some utilities with existing WebKit2 API tests (that we run as well).

This commit adds just a skeleton for the paintToCurrentGLContext test
and removes the previous infrastructure.

I explicitly opted to not use neither a Test Fixture, to make the
tests simpler and more self contained; nor PlatformWebView, since the
focus here is to test our WebView API, it is good to keep its code
usage visible in these tests.

Reviewed by Luciano Wolf.