Make LayerTreeRenderer initialize as active
So we do not need a first paint call to activate it. The
LayerTreeRenderer sets it to false when purging resources. This part
of the code should be improved in upstream. This is based in kbalazs
patch for QtWebKit in r129782.

This will be useful to proper implement our API tests.

Reviewed by Jesus Sanchez-Palencia.