Use our own GLContext for GLX
We need to store the Display and the GLXDrawable to proper make the
context current for an GLContext that adopts the current context in
GLX. This commit create our own version instead of tweaking the
existing GLContextGLX, that would need some changes (or become ugly)
to cover our case.

I removed USE(GLX) since we don't use upstream GLX code for now.

This should fix the ASSERT() when running MiniBrowser in Debug mode.

Reviewed by Hugo Parente Lima.