TouchMocker should not always trigger an updateDisplay on MiniBrowser
Hugo noticed slowlyness on his machine after that TouchMocker's patch,
so I've optimized to only update the screen when necessary. That means,
we should only update the screen whenever we know a touch was really added,
moved or removed, and not when we just don't want to move the event forward.

Reviewed by: Caio Marcelo