Callback doneWithGestureEvent should use NativeWebGestureEvent
I had to touch public interface to use NativeWebGestureEvent instead.
This callback is also working on Qt. I've moved the event creation to
WebEventFactoryQt to keep code standards. On mac, just renamed the type
as they don't use it yet (so it's ok to forward declarate it). On nix, we
expose the callback as virtual but we offer a no-op implementation of it.

Reviewed by: Caio Marcelo