Add webkitpy layout-test's unittests for nix
Our port is very special as we only have support for wk2.
So I had to fix in a few places to avoid redundant search paths
(we should not re-add the same path name to that list). Our wk2 portname
is the same (i.e. "nix") instead of the current convention ("port-wk2").
There was also a test for timeout but it was assuming a port always
have wk1 support which is not true when nix is around.

Reviewed by: Lauro Moura