Add a GestureRecognizer to MiniBrowser.
Instead of the TouchMocker doing the job of detecting
when a tap occured it'll just generate the touch events and
send them to the WebView. Any touch events not handled by
webkit are then fed to the GestureRecognizer. It's slightly
based on the QtGestureRecognizers but simplified for
our use case.

Currently SingleTap and Panning have callbacks implemented.

Reviewed by: Caio Marcelo de Oliveira Filho