Gamepad API for Nix
Nix now has a Nix::Platform namespace, living in
Source/Platform/Nix. The entities being provided right now
by Nix::Platform are a GamepadDevice struct and an abstract GamepadClient.
This client has to be set by the application using Nix, and there is a
getGamepad() to be used by WebCore (WebCore/platform/nix/GamepasNix.cpp)
when sampling Gamepads for the Gamepad API implementation.

Nix::Platform is a just a way to hook device api implementations to
directly to the application, and all the life cycle of the needed
data structures should be handle by it.

Reviewed by: Caio Marcelo