Add a way to mock scaling in MiniBrowser
The new function provided in MiniBrowser to scale centering in a
certain point of the screen. The idea is that contents below that
point will remain in the same "window" position after scale is
applied.

We use SHIFT and mouse wheel to get this behavior. The plan is to
use this in the future together with Pinch handling.

Reviewed by Hugo Parente Lima.