Enable 3D Transforms support in Nix
The implementation is already there, you can check in MiniBrowser
https://www.webkit.org/blog-files/3d-transforms/poster-circle.html
to have an example.

New tests run after this, and I added some expectations. Some of those
tests only make sense as pixel tests, so we'll re-evaluate them later
when we add support for taking the snapshots of WTR in the UI process.
Others (like the hit testing) work programatically, so they are OK.

Reviewed by Rafael Brandao.