Nix events (x,y) values now refer to contents (web page) coordinates.
Such event coordinates previously referred to the GL context.

GestureRecognizer::handlePanning method passes the touch point where
the panning started, instead of the previous touch in the movement
sequence.

In the near future the GestureRecognizer will work with the global
positions to decide which gesture is meant by the user, since global
coordinates provide a more realistic way to evaluate user movements.
And the recognizer will call it's client with data calculated from
the contents position.

Reviewed by: Caio Marcelo de Oliveira Filho