Fix scaling when mocking pinch
The bug was seen when starting a pinch in a page current scaled at
~0.8. The information given by us from GestureRecognizer is the
distance ratio, considering 1.0 the ratio when the pinch starts. We
need to keep the scale at start of the gesture as well, so we can
calculate the new scale given a new ratio.

Since we already have the scale information in gesture recognizer,
we put the calculations of new scale inside it as well. The rationale
here was to avoid keeping state in the MiniBrowser to track this and
let recognizer do the job for us.

This patch also changes scaleAtPoint() to take the actual scale as
argument instead of the scale ratio. This avoided some unnecessary
roundtrips (dividing by m_webView->scale() then multiplying by it
again).

Reviewed by Marcelo Lira.