Don't use a bunch of shared objects on Nix, bundle everything inside WebKitNix.so.
The big picture:

- WTF, JavaScriptCore, WebCore, WebCoreTestSupport and Platform are compiled as static
  libraries.
- Platform library doesn't link with any WebKit module, and it shouldn't.
- WebKitNix links to all these static libraries, but using --whole-archive link flag
  to ensure all exported functions inside these static libraries will be inside the
  final shared object.
- Added Nix functions to access WebCoreSupport functions, so every access to
  WebCoreTestSupport must go via WebKitNix library.

To avoid future problems, with objects instanciated in one shared library and deleted
in another one WTF::fastAlloc and WTF::fastFree aren't used anymore, other ports also
does the same (Qt and GTK).

In the end, just WebKitNix.so and the include headers must be installed on the system.

Reviewed by: Caio Marcelo de Oliveira Filho