Adapt the window size in accordance with device defined to MiniBrowser.
The window size should have the dimensions of the device defined to Mini Browser, unless a size has been chosen to the window.
The default width and height of window was changed to 1024x768.
A table with the information of devices (width, height and user agent) was added for easier access.

Reviewed by: Caio Marcelo de Oliveira Filho