Revamp alpha handling in ImageDiff.
Cairo's png writing functions checks if whether the given surface
is opaque, forcing a non-alpha png output. This behavior was breaking
with the previous ImageDiff behavior, that checked if the test image
had the same number of channels and alpha of the reference image.

This commit changes ImageDiff to multiply the pixel value by the alpha
and then comparing only the RGB values.

Also:

Fix reading of png images plus better error log.

Was causing errors when reading the WTR output due to
ignoring IEND chunks. The call to png_read_end is enough
to fix.

Reviewed by: Caio Marcelo de Oliveira Filho