Fix content being blurry with different scale values
Besides being correct, applying the translation after the scale
would mean that the translation is affected by the previous scale,
and in some cases the resulting offset would be non-integer number
which was causing the content being blurry when scrolled down.

So we switch the order between the translate and scale, doing the
translation first. But this translation already takes the scale into
account, so it would produce the exactly same resulting matrix as before.
The trick here is to round this offset at this point, and then we apply
the scale afterwards. So now the matrix we produce is slightly different
than before, our offset will never be non-integer anymore.

This patch is based on http://trac.webkit.org/changeset/136000.

Reviewed by: Luciano Wolf