Reimplement determine_full_port_name for nix.
This function is used to determine port_names based on values
from the host and option arguments, e. g. operating system versions and
qt versions.

The default behavior from base.py is to add the wk2 suffix for WK2-based
ports. This commit reimplements it for nix to return only 'nix'.

Reviewed by: Rafael Brandão