Fix memory leak when generating snapshots
Our return type is WKRetainPtr<WKImage>. Returning a bare WKImage will
cause the creation of a new WKRetainPtr which will retain the WKImage
once more. Since the function doesn't explicit release, we leak.

In those cases the approach is to use adoptWK() that will create a new
WKRetainPtr adopting the WKImage (not retaining again). This way the
ownership is correctly given to the caller function without leaking.

Reviewed by Lauro Neto.