Use TransformationMatrix in our WebView implementation
We used to use cairo_matrix_t as part of web view API, and it was convenient to
use cairo_matrix functions to do calculations. Now we don't have cairo_matrix_t in
the API anymore, so there's no reason to perform calculations using cairo, we can
simply use TransformationMatrix.

Reviewed by Hugo Parente Lima.