Unreviewed gardening after merge to make bot green.
Some of the compositing tests are crashing, it may be related
with the bug https://webkit.org/b/104506 so marking them as expected
to crash for now. Other regressions needs more investigation, but
there was some compositing patches on this merge that may change
our results in tests or even highlight real bugs.

Reviewed by: Trust Me