Disable gesture recognizer of MiniBrowser when page have touch handler.
Pages are able to handle touch events, using the ontouchstart, ontouchmove and ontouchend listeners.
But if they don't treat some of these events, the gesture recognizer couldn't know what gesture to do, or if the gesture was already been done by the page.

So if the page consume any touch event sent, the gesture recognizer of the MiniBrowser need be disabled to avoid mistakes.
At the beginning of the loading of a new page, the gesture recognizer must be enabled again, this was done at didStartProgress callback.

Reviewed by: Caio Marcelo de Oliveira Filho, Rafael Brandão