Gesture recognizer should be resetted when touch is handled by WebProcess
In the past, we could hit some unexpected states on gesture recognizer when the page
was dealing with a subset of {touchstart,touchmove,touchend}. At that time, we've decided to
shut down in gesture recognizer  in a page that would manage any of them, but this is tricky to
maintain and "didStartProgress" does not always track every real "page change" or any part
of it (via ajax for example). So once we get into the dead state of not recognizing any gesture,
we would never get a single tap and never do any navigation.

An easier solution is just to reset the state of the recognizer whenever the page handles
any touch event, since recognizer might be damaged by untracked events. Now, at the initial state,
we admit the possibility to get things like "move" or "end" without a preceding "start", but we're
just not moving forward in the state machine.

Reviewed by: Luciano Wolf
             Tullio Lucena