Long tap should trigger a context menu
Once we get the first touch start from initial state, we move to
state "single tap pressed" and start the long tap timer. If we get
any touch before the timer finishes, then we don't have long tap,
so the timer is stopped and the state machine moves forward like
before. Otherwise, if no touch happens before the timer ends, then
we have a long tap, and reset the state machine again.

In order to request a context menu, we use a synthetic mouse event
to simulate a right click on that point.

Reviewed by: Jesus Sanchez-Palencia
             Hugo Parente Lima