Doesn't depend on X11 for EGL.
When using EGL Nix uses the OpenGL abstraction of GLContext class, when
using GLX it uses the abstraction of GLPlatformContext.

At the moment we can't use the same code path for both because GLPlatformContext
depends on X11 for EGL, on the other hand Graphics Surface isn't enabled for
GLContext code path.

Reviewed by: Jesus Sanchez-Palencia